pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tasty_api"
version = "0.0.1"
authors = [
    {name="Kian Khadempour"}
]
description = "Typed Python wrapper for the Tasty API on RapidAPI."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Private :: Do Not Upload",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed"
]
license = {file = "LICENSE"}
dependencies = [
    "pycountry==23.12.11",
    "python-iso639==2024.4.27"
]

[project.urls]
Repository = "https://github.com/KianKhadempour/TastyAPI.git"
Issues = "https://github.com/KianKhadempour/TastyAPI/issues"

[tool.ruff.lint]
# Disable fix for unused imports (`F401`).
unfixable = ["F401"]